feat: introduce --no-follow-logs cli flag
