+++ internal/task/consts.go
@@ -1,0 +1,8 @@
+package task
+
+const (
+	labelKeyTaskName     = "k6ctl/task"
+	scriptsVolumeName    = "k6-scripts"
+	containerScriptsPath = "/scripts"
+	containerNameRunner  = "k6-runner"
+)
